requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
markdownify==1.2.2
packaging==26.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
zstandard==0.25.0
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.2